pyproject.toml
[project]
name = "codebase-intelligence-hub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"

requires-python = ">=3.11"
dependencies = [
    "langchain-core>=1.2.19",
    "langchain-openai>=1.1.11",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-chroma>=0.1.0",
    "chromadb>=0.4.0",
    "openai>=2.28.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "langserve>=0.2.0",
    "unstructured>=0.21.5",
    "tqdm>=4.67.3",
    "rank-bm25>=0.2.2",
    "sse-starlette>=1.6.5",
    "langchain-experimental>=0.3.0",
    "sentence-transformers>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pyright>=1.1.0",
    "bandit>=1.7.0",
    "semgrep>=1.45.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "radon>=6.0.0",
    "pylint>=3.0.0",
    "pip-audit>=2.6.0",
]

docker = [
    "docker>=6.0.0",
]
